Fix for ImageInterface.values
